requirement.txt
numpy==2.2.2
opencv-python-headless==4.12.0.88
pyzbar==0.1.9
fpdf2==2.8.5
matplotlib==3.10.6